bus-ified all signals from sensors
Allows to access inputs in matlab function blocks as structures. Buses
are defined in the newly tracked "bus_definitions.m" file. If a change
is made in the layout of the model(regarding those buses), the changes
shall be mirrored in the code of "bus_definitions.m"

using buses allows for more flexibility in the amendment of the model.
The code is not dependent of the signals position and more readable.