Loop-closing v0.9
Not functional yet because of some Simulink bugs

Modified the low level controller block in "Autonomous Driving". Added
low level control loops. Only shaping controllers for the moment
(feedforward). For the steering the mapping is approximative so a real
feedback is neeed. It will also be neeed in Torque steering mode (only
angle steering for the moment).

Restructured a bit the bus definition file to allow for more flexibility
in the order of signals (use of indice i instead of hard numbers).

Added switches for autonomous behaviour/driver. Still need to define the
switch logic.